refactor: move auth screen to an auth folder
