+++ src/screens/public/auth/components/sign-in-form.component.tsx
@@ -3,14 +3,14 @@ import { useForm } from 'react-hook-form';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { z } from 'zod';
 
-import { SignInService } from '../../../modules/user/services/sign-in.service';
-import { FindOneUserService } from '../../../modules/user/services/find-one-user.service';
+import { SignInService } from '../../../../modules/user/services/sign-in.service';
+import { FindOneUserService } from '../../../../modules/user/services/find-one-user.service';
 
-import { TextField, Button } from '../../../common/components';
+import { TextField, Button } from '../../../../common/components';
 
-import { useLoading } from '../../../common/hooks/use-loding.hook';
+import { useLoading } from '../../../../common/hooks/use-loding.hook';
 
-import { EFirebaseErrors } from '../../../infra/firebase/enums/firebase-errors.enum';
+import { EFirebaseErrors } from '../../../../infra/firebase/enums/firebase-errors.enum';
 
 import type { FC } from 'react';
 import type { SubmitHandler } from 'react-hook-form';
